Explain pip's "editable mode"
The follow up to Clemens' commit. It adds a section to the (now
capitalized) README.md explaining how to install python packages which
have to pick up live updates without reinstallation. Doing it this ways
is roughly as much work as using submodules, but keeps the source tree
clean.
Note I have also added a Wiki entry similar to this and the `pvlib` devs
also mention this option in their README, so depending on your taste
this could be removed if we deem the README the wrong place for such an
explanation.