Make 'powerplants.py's '__main__' stanza runnable
This necessitated a few changes. First a few imports had to be added as
the code silently assumed them to be present.
After that the code assumed two variables to be present which where
never initialized. Fixing this was a bit more complex.

In order to fix the first one, I made it an additional keyword argument
to 'feedin'. This means 'feedin' is no longer a property but can be
treated as a proper member function. Unfortunately it also means that
the function got a little bit more complex. One can now use arbitrary
keyword arguments which will override 'Photovoltaic' attributes if their
keys match the attribute's name.

The last one was the 'site' parameter. It looked like this one should be
an attribute of a 'Photovoltaic' instance to me, so I modified
'Photovoltaic's initializer to add a reference to the 'Photovoltaic'
instance to the model. That way one can get at 'site' attribute from
within the model. Might be overly complicated, but for now it makes
sense. (To me at least.) We'll see.

The calls to 'Photovoltaic's constructor and the to 'feedin' have then
been updated accordingly in the example.