Fix error caused by lists as default arguments
It's a common mistake and one that happened to me multiple times. Using a
mutable list as a default argument value in a function definition means
that the list persists across function calls. This means that if the
default argument happens to kick in, every function call will use the
SAME list which is almost never intended.