Add example of how to use the ORM
The db 'module' added in this patch does not contain any functionality
but for now just acts as a script demonstrating how SQLAlchemy could be
used. In order to run it one would have to have a fitting database and
change the connection string accordingly.
A good way to move on would probably be to create a mock up CSV file
with few lines of imaginary data, read that into an in memory database
and let the example work on that. This setup could then also be used to
write tests for the interaction between ORM and 'Components'.