Add solve_opt_model() and replace pyomo.Set() objects by list/dicts
with solve_opt_model() a created pyomo instance can be solved. Note: solver
must be installed.

I looked at building time of a pyomo model, which was really long, so I
tried to replace set object of pyomo be simple lists for simple sets and
dicts for indexed sets. It turned out, that building the model was over ten
times faster. (hope this is right)