Restructure 'Entity' inheritance tree packaging
In order to not have everything in one big file, the classes in the
inheritance hierarchy rooted at 'Entity' have been split into packages
such that each class is contained in a module named after its parent in
the inheritance tree and (sub)package paths mirror the inheritance tree.
An added test ensures that all modules load properly and that the
inheritance chain is intact.