Refactor `EnergySystem#connect`
There are a few things happening here. First, the `classtype` argument
is no longer a string, but an actual type object which can directly be
used to instantiate a transport. The argument has been renamed
accordingly. The log message now also uses the argument to give details
about the type of transport created.
The argument has then been used to get rid of the conditional assignment
as well as the inner function and do transport creation and the
assignment to the `connections` list in one go inside a loop.
As a last step the error message displayed when a class which differs
from `oemof.core.network.entities.components.transports.Simple` is used
has been elaborated a little bit.