Change assemblers to single dispatch functions
That means they are no longer instance methods of an `OptimizationModel`
but module functions. On the upside we can get rid of the `lower_name`
class attribute. No ugly string/getattr magic is necessary anymore,
since assemblers are now selected on the basis of the type of their
first argument (or they can be directly selected if the target type is
known).