Update `OptimizationModel` attribute access
The attributes accessed in the changed code came into existence due to
the use of `add_component` inherent to Pyomo's `ConcreteModel`s. The
attribute names where derived from the `lower_name` of the corresponding
object types/classes and this class attribute is gone now. Instead, the
type's string representation is used now. Since this string is usually
not a valid attribute name, it has to be accessed through `getattr`
instead of using normal access methods.
We should probably find a different (read: better) way of doing this,
since strings are not exactly a robust way of doing this, but for now it
works.