Back up stuff for later use
This is some work in progress that I'm going to need later when it's
time to add some features to the optimization results dictionary. The
code `postprocessing` module is going to be removed in the next few
commits anyways and I'll need the code somewhere else.
When the time comes, I'll check out the code and copy it to where it
belongs then.