Muliple fixes on `registry` attribute dosctring
The apostrophe plus "s" construct in line one was wrong and is now
converted to a proper plural with correct hyperlinking.
Using the `:const:` role was also wrong, as `None` is not a variable
that should not be changed.
Last but not least, the `... on construction.` addendum explains what is
happening a bit more precisely.