Put computed values into results object attributes
I don't think users will expect to get `pyomo` `_VarData` objects as
results. Therefor I changed it so that the computed numbers are put into
the attributes.
The only exception is the `objective` attribute which still holds a
`SimpleObjective` object where the value has to be explicitly requested.
But since all of this is in line with the existing documentation, that's
what I went with.