Test whether sinks appear in results dataframe
For some reason `Sinks` currently do not appear in the results
`DataFrame` after it's generated from the results dict, IF the `Bus` to
which the `Sink` is attached has dual values.
This commit adds a test that checks whether the `Sink` is present in
both cases, and currently fails properly for the case with `duals`
present.