Refactor results dataframe for `Bus`es
There was a lot of repetition in the `if` branches that pulled the
results (where a `Bus` acts as the top level key) out of the results
`dict` and into the results `DataFrame`. I condensed this to minimize
repetition.
The actual reason for doing this was fixing the bug where `Sink`s where
missing on `Bus`es with duals in the `DataFrame`. The short solution
would've been to just add an `else` branch which was missing, but doing
it this way is better IMHO.