Add a failing test for 'investment' defaults
There are lines in a previous test which would make this test pass. As a
consequence those lines have been refactored to proper setup/teardown
code.
The test fails because of interactions with other tests which set values
in a dictionary which is a class attribute without resetting the
values. In order to fix this those tests have to get proper
setup/teardown coded too.