Refactor the PostHeating class and its solph functions
The PostHeating class was created to describe a instantaneous
flow heater. The new idea is describe a more generic class that
describes a general transformer that has two inputs and one output.
This transformer can represent a heat pump, an instant flow heater
or any other component with two inputs and one output.