Explicitly set test precondition
This test worked before, because the condition it relied on was set in a
different test that ran before this one. This doesn't work anymore as I
added setup/teardown methods which make each test run under default
conditions to the test suite. That's why the tests condition now has to
be set at the test's start.