State requiredness in parameter explanation
When using Sphinx with the numpydoc docstring convention, the top level
directly below the `Other  Parameters` section apparently is expected to
contain lines of the form `parameter name : type`. Doing it differently
makes documentation look weird. For this reason this patch moves the
specification of which value of the `method` parameter requires which of
the `Other Parameters` from the top level into each individual
parameter's explanation.
Note that one could drop (maybe more than one) level of indentation but
the patch doesn't do this to make the diff more readable.