Reference `uid` as an attribute
It's not declared as an attribute in the `Entity` class, so this
reference doesn't resolve to anything. But doing it this way future
proofs the documentation for when/if `uid` is added as a documented
attribute of `Entity`. Additionally, the attribute is already referenced
this way previously in the same sentence and that patch streamlines how
it's displayed in the documentation.