Locate "lp_files" relative to test file
Previously, the directory containing the "lp-files" against which tests
compare the generated ones where located relying on the current working
directory so the tests had to be executed from the repository root.
Now these files are located relative to the file containing the tests so
the tests can be executed from other places too.