Ignore trailing whitespace when comparing lp-files
Whitespace at the end of lines shouldn't change the semantics of the
files so it should be ok to just ignore it. More importantly, for some
reason pyomo seems to be unable to make up it's mind on how much
whitespace to put at the end of some lines so tests where failing when
they shouldn't.
Note that ignoring whitespace also ignores trailing newlines so I had to
change the way the diff is output by adding newlines back and specifying
an explicit empty `lineterm` argument to `unified_diff`.