Ignore lp-file's "objective" part in certain tests
It seems like pyomo changed the way placeholder objective expressions
are generated between versions. As this is of no importance for some of
the tests, these generated lines are now ignored there using the regular
expression "^objective.*(?=s\.t\.)", which matches a line starting with
the word "objective" and everything after that until the string "s.t.".
The `re.MULTILINE` flag enables "^" to match at the beginning of a line
while the `re.DOTALL` flag means that the dot, i.e. ".", also matches
newlines, i.e. "\n".