Fix accessing `outputs` of a node having none
Nodes without any outputs are initially not present as keys in the
`_flows` dictionary. Returning the default is handled in the
`__getitem__` method of `_Edges` instances, i.e. by using `flow[node]`,
which was bypassed by directly accessing `_flows`.
The fix: just don't do it and rely on the correct behaviour of
`self.__getitem__`.
Should've done this from the get go.
The upside: now we have a test for this.