Defer groups update until next access
That way the actual regrouping is only done when the `groups` property
is accessed again. If I'm not totally mistaken, this should make sure
that the grouping functions get to act on fully constructed
nodes/enttities. Unless you're accessing any owning energy systems
`groups` property during `Entity`/`Node` construction, that is. But one
shouldn't do this anyway. A `Node`/`Entity` should not care about the
energy systems it is located in.