Consider both directions when grouping flows
Only considering outgoing flows is correct when the energy system is
completely constructed, but when not all nodes and their corresponding
flows are known, it means we might miss some flows.
For this reason it's better to iterate over both directions and make
sure we are not duplicating flows by storing them in a `set`.