Develop approach further, fix weakref error
In order to fix the setting of node inputs and outputs buses are
created to replace the labels as dictionary keys. Nodes are created with
a generated subset per node for each unique label. The dictionary
containing key word arguments for initializing a node can be built
based on the node type.