Move grouping code to it's own module
This slightly more than just a cut & paste operation, as I also
implemented using lists of group keys instead of `MultipleGroups`
instances. The other option would have been to move `MultipleGroups`
into the new module and then deprecating it and moving it back to
`energy_system` once the additional feature is done.
Way to complicated.
With the current way the outside interface doesn't change (much) and
it's easy to add a deprecation note later on.