Use undocumented object to set `inputs`/`outputs`
Since just `update`ing the dictionaries doesn't work, we have to resort
to using an undocumented object to set a `Node`'s `inputs` and `outputs`
after it's creation.
The `flow` object in `oemof.network` shouldn't be considered part of the
stable API of `oemof.network` but it works for now.