Fix mistakes made when resolving merge conflicts
For the `_groups` case I simply chose the wrong option to resolve the
merge. The `_groups` attribute is the right choice, as `groups` being a
property backed by the internal `_groups` is the current state of
affairs.

For the `import ... as OM`, I actually re-introduced a deleted line when
resolving the merge, which is wrong. In order to fix this, I only had to
remove the line again, but I also future proved the moved import for
usage with the new solph.