Fix mistake from commit 2cc2c47
That commit moved an import under the assumption that it was dealing
with a circular import. That was wrong. The actual problem probably was
that the first import was gone due to the refactoring process.
Doing a conditional import seems like it's the right decision. At least
the test behaviour doesn't change.