Revert "Move grouping from __init__ to method"
This reverts commit 00cfff5f6a1464977a5f90fe391986fb738a67c3.
The idea to re-group an energy systems contents after they have been
added to the energy system is a bigger can of worms and if we go through
with adding this functionality it's better to do this on a separate,
dedicated branch.
For now, creating the energy system AFTER all the nodes have been
created should be a functional workaround.