Add failing test case
Although [0] already fails, this one is a pretty course test. The new
test specifically tests the actual failure reason, namely that an
investment flow isn't grouped properly.
In proper TDD/TFD fashion, this test is added first, while it still
fails, so that we can make sure that the next commit actually makes the
test pass.

[0]: tests/constraint_tests.py:Constraint_Tests.test_fixed_source_invest_sink