Fix `InvestmentFlow` grouping misbehaviour
Turns out that both, in- and outflows are collected in the group, but
they are not checked to generate the key. So in the case of outflows
being connected after instantiation of their outgoing node, they where
not collected in the group.
The fix was easy. Finding this error was the hard part.