Fix (probably) misspelled attribute access
More specifically change

  `m.variable_costs[i, o][t]`

to

  `m.flows[i, o].variable_costs[t]`.

The error says that `OperationalModel` (which is the type of `m`) has
not `variable_costs` attribute and since the if's condition makes sure
the attribute exists on `m.flows[i, o]` (and is not `None`) I'm guessing
this is what was actually meant here.

This fixes the attribute errors but leads to new ones, of which we'll
fix at least one in the next commits.