Stop modifying groups in place
It doesn't work when storing groups under multiple keys in one go, as
the group than is just an alias that is basically the same everywhere,
which it shouldn't be.
So subsequent `merge`s might modify a group under a key that shouldn't
actually be modified.
Aliasing. It always bites one at one point.