Fix adding custom `CONSTRAINT_GROUPS`
Using `extend` changes the class variable in place, which is not the
desired behaviour as it changes the `CONSTRAINT_GROUPS` of all future
`OperationalModels`. What we want is a new temporary list, which is
only used for this model. This is exactly what we get when using `+`
instead of extend.