Conditionally change docstrings to not be doctests
The doctest of `graph` fails if the optional packages are not installed.
On import, the docstring of that function is now changed to not look
like a doctest anymore, if the test conditions are not met.
This is done by dropping one '>' or '.' from the beginning of lines
looking like interactive python sessions. That way they still look a bit
like python sessions, but are not treated as doctests anymore.