Refactor `oemof.solph.Flow` initialization
Setting the default values was interspersed with other code and it was
done manually without variables storing which attributes are to be
handled specially. Now the code is in one place, in front of the code
which is concerned with other, more detailed checks.
Additionally it's now easier to figure out which attributes should be
handled specially.