Enable modifying a `Node`'s inputs and outputs
Outputs worked out of the box, but input modifications had to be
rewired to do the symmetric output modification instead. This is done
using a custom dictionary that stores the target of the inputs we're
interested in and uses that to map actions on `target.inputs[source]` to
the corresponding actions on `source.outputs[target]`.