Revert adapting tests
Part of this reverts commit 45fedc89290f63271c9a073977257dfd29fb0f90.
We're fixing node ordering instead, so having the tests rely on the
ordering implicitly tests that it works as expected and should therefore
be reverted to the old behaviour of statically comparing the results at
fixed locations.
This is not an exact revert though, as the order is now different from
what it previously was. In order to clarify, how the order is obtained,
I wrote the keys used in sorting as comments besides their values.