Add two auxiliary functions
The first function, i.e. `parse`, takes care of JSON strings where keys
are not properly quoted which is necessary because we wanted to reduce
the amount of quotes people have to write when manually creating CSV
files for the datapackages.
The second one, `listify`, creates infinite lists of stuff which is not
boxed in a list. This is for occasions where multiple flows share the
same parameters. In those cases you can only specify them once and they
are automatically boxed in a list.