Use a dictionary to store resources
That way it's easy to return all of them. This is also more extensible
should the need to read in other resources arise and doing it this way
also enables me to eliminate the duplicate code use to read in "hubs"
and "components" which I'll do in the next commit.