Make deserialized `Node`s serializable
Serialization is only possible if initialization arguments of a `Node`
are saved in it's `._state` attribute. This attribute was only set
during `__init__` and not during `__setstate__`. Since deserialization
bypasses `__init__`, `Node` created be deserialization did not have a
`_state` attribute and therefore could not be serialized again.
This commit moves setting the `_state` attribute to the `__setstate__`
function, where it belongs, which fixes the issue explained above.