Resolve object references in component parameters
This has to happen before the objects are instantiated, as the
constructor otherwise creates conversion factors for those "missing"
string inputs and outputs. This meant that the function resolving object
references had to be defined prior to component creation, leading to a
reordering of the code.
Also, it now seems unnecessary to resolve object references across all
resources. Let's see if that's correct.