Streamline and fix formula symbols and explanations
The condensing efficiency is related to the other efficiencies so its
symbol was changed to be in line with those efficiencies. The variable
that is used for it in the code was also named incorrectly, which gets
fixed in this commit. Lastly, $I_{mfl}$ can be expressed in terms of the
three efficiencies, which is done now too. The explanations get adjusted
accordingly.