Two TODOs done, one more acquired
Generating errors for missing required parameters is now done in a
uniform way in `Facade`s. This means we don't have to rely unnecessary
class variables AND I followed up on my own comment on how to handle
things slightly better.
Unfortunately we might still hit an error case before the objects are
actually constructed, but this has nothing to do with the objects being
construct but only with how we read the data package. Therefore claiming
that the attribute is required for the object would be slightly
misleading. As it is important to point to the correct reason for an
error, I added a TODO so that I know that I should generate an
appropriate error message eventually.