Fix `.travis.yml`
Although `datapackage` is an optional dependency, the module using it
still gets run by `nosetests --with-doctest` since that tries to load
every module in order to discover doctests. That's why the travis
configuration has to be changed to install the optional dependencies
too, when installing the package to test it.