Remove docstring for nonexistent parameter
The functionality the parameter documentation promised was never
implemented. Since no one tried to use it, is also seems like it's not
necessary to implement it. On top of that the parameter's name now
collides with an attribute that will be added during this refactoring so
it's a good time to remove that outdated piece of documentation.