Add the `Edge` class
A third building block of an energy system network. The fact that it's a
subclass of `Node` might look weird if you're coming from a graph
theoretic background but we'll either change that name, or explain it in
the documentation later.
The short version is: `Node`s are nodes in a graph theoretical sense,
Let's call the graph, they belong to G. G is, or more specifically its
edges are, never actually constructed. Instead we derive our network N,
which is also a graph, from G by treating certain nodes in G as nodes in
N and certain other nodes in G as edges in N. This is why all the
building blocks of G are `Node`s.