Remove flow values from `Edge.Label`
I thought having those values on an `Edge`'s `label` would be good for
debugging purposes, but as these values can also be rather large
sequences, it's probably not a good idea to always have them in the
string representation. Also, having the flow values be part of
the `label` means that parallel edges are possible in a single energy
system, but only if their initial flow values are not equal, due to the
way groupings currently work. On top of that, these initial flow values
would be part of the key for an `Edge`'s singleton group. Whether we
want to have parallel edges or whether we want to differentiate
different flows between the same `Node`'s in a different way actually is
a non-trivial design decision that we have to think about separately.
And having the initial flow values be part of a string that is used as a
key in a dictionary definitely is something we do not want.
Hence the parameter should be removed from an `Edge`'s `label`.