Change `Outputs` to work without `flow` or `Edges`
This allows us to get rid of a lot of code. Since `Inputs` is already
changed to relegate to instances of this class, all that has to be taken
care of is that the targets `_in_edges` are set and deleted appropriately.
Subclassing from `UserDict` instead of `MutableMapping` is done in order
to not have to provide trivial implementations for some of the methods
that are abstract in `MutableMapping`.