Allow delaying registration of a `Node`
This is done by setting a specific attribute on the `Node` during
construction. Note that this means registration has to be done
explicitly at a certain point AND the attribute has to be unset before
attempting to do so. I didn't really document this, because I'm not yet
set on the interface for this, but I made a big TODO reminding me to
stabilize the interface for this and to document it properly.
Also, since registration starts to become a bit complicated with this,
it seems like a good idea to factor it out into its own method, which I
took the liberty to do with this commit, too.