Allow constructing an `Edge` without `input` or `output`
This commit is a bit more complex, as a few things have to be taken care
of.
One thing is, that we obviously don't want an `Edge` without `input` or
`output` to be registered and consequently grouped, as an `Edge`'s
label is generated from `input` and `output` and having one of those set
to `None` easily clashes. This is why we have to delay registration and
fire it off manually, once both, `input` and `output` are set to non
`None` values. Conveniently, the checks to do this an be place in the
setters for the two properties, which have to be written anyway, as the
values would not be settable otherwise.
Next, we can not simply set the `input` and/or `output` component of the
`Edge`'s `label`, as the `label` is a `namedtuple` which is immutable,
and for good reason[0]. That means we have to make `label`s assignable.
That's why the `label` property now has a setter. This means, whenever
`input` or `output` gets assigned, a new namedtuple can be created and
assigned.
The last thing to take care of now is to also update the `Edge`'s
endpoints' `inputs` and `outputs` appropriately, once `input` and
`output` are both different from `None`.
There are a few more potential caveats, but those get a reminder in the
form of a TODO.

[0]: The reason being, that we want our `label`s to be hashable and
     hashable "containers" with a `hash` implementation that depends on
     their contents should NOT ALLOW their contents to be changed.