Make `solph.Flow` a subclass of `Edge`
As the interface for `Flow` didn't specify it having `input` and
`output` it is used without setting those on construction everywhere and
instead relying on the `Flow` being set via a `Node`'s `inputs` and
`outputs` construction parameters and attributes.
This is the reason we have to have the `from_object` constructor for
`Edge`s, as this helps us to not create `Edge`s containing `Flow`s as
`values`. Instead, the `Flow` is taken as is and its `input` and
`output` are set correctly when a `Node` is constructed.