Fix registration tests
They where relying on the order in which `Node`s/`Entity`s got
registered. This is to strict. The fact that they got registered should
suffice. This got uncovered because now `Edge`s get created, and, now
that they are subclasses of `Node` registered implicitly which messes up
the ordering.