Change processing tests to use a smaller timeindex
There's really no reason to test 8760 values. Especially not, if they
are mostly the same. Tests should be small, minimal in the best case.
What's more important is that this leads to unreadable output on test
failure. Which I got. And without this change I would have had a really
hard way of figuring out why the test failed.
So I changed the timeindex's to length the minimal amount of two for the
two different values used.