Get rid of customized pickling
The tests for an energy systems `dump` and `restore` functionality kept
failing with the new explicit `Edge` class. The reason for these fails
where in the code customizing how `Node`s are pickled. When researching
this, I figured out that due to the refactoring, the original reason for
having to customize how `Node`s are pickled is gone now, since there is
no `_Edge` class anymore. This means `__getstate__` and `__setstate__`,
which customize how objects are pickled, are no longer necessary and the
`__setstate__` code can be moved into `__init__` again.
This gets rid of a huge TODO, removes obscure "magic" from the core and
enables us to remove some `__slots__` entries. The `_state` entrie was
only necessary because `__getstate__` and `__setstate__` depended upon
it for storage. The `__weakref__` entry was already obsolete after
`_Edge` got refactored to `Edge`, but wasn't removed because I probably
didn't realize that at the time.