Fix merge hickup
I reduced the number of time steps tested to speed up the tests, but it
seems that during merging the tests got their old values back while the
time steps retained their reduced value. This commit goes back to the
non reduced number of time steps to fix the tests.