Stop using sets of strings
Using sets was an oversight anyway. Simply having the strings there
makes way more sense for now. Should we figure out that we need multiple
equations per key, we can always change to using a collection of choice.

I also flipped some equations to have the shorter term on the left hand
side, so that aligning the equations is easier when rendering them.