Catch the `DeprecationWarning` and test it
The `catch_warnings` context manager manipulates the global warning
filter. Everything run with this context is treated specially. When used
with `record=True`, warnings are not printed to stderr and are instead
added to a list, which is called `warnings` here. That way one can test
how many and which warnings are issued. The warning objects are not
added to the list directly, but are instead wrapped in another object.
The warning objects can be accessed via the wrapper's `message`
attribute.