Fix two bugs in handling immutable mapping groups
First, filter is used wrongly. Since it might also be `None`, and will
be so by default, it has to be plugged into the `filter` built-in here.
Also, we can only be sure that a `Mapping` instance can be constructed
from another `Mapping` which is why the argument generator has to be
wrapped in a `dict`.