Make summed_min/max in Flow flexible
In #573, it has been suggested to allow to split the total time into
"subperiods". This commit suggests to convert scalar summed_min
and summed_max attributes to dicts instead of enforcing dicts with
just one entry. This way, internally, a one-entry dict is used.

(The alternative approach which keeps the scalars and only uses
subperiods when defined might be faster but harder to maintain.)