Partially trigger #576 in a test
The `investment_flow_grouping` looks at `Flow`s and their endpoints and
puts them in a group if the `Flow` has a non-falsy `investment`
attribute. According to #576, this doesn't happen if the objects
connected by the `Flow` are already `add`ed to the energy system
containing the `investment_flow_grouping`, which is tested for here.
The case that is specifically tested here is the one, where `groups` is
not accessed until the `Flow` and the objects it is connecting are
`add`ed.