Make energy system `groups` fully lazy
This has benefit that it:

  * streamlines energy system initialization code,
  * gets rid of private functions like `_add` and `_regroup` and
  * adds more options on how to construct `Node`s and `add` them to an
    energy system.

See also f826a5d07b5466d990302341661ae2a64caea5b8 for more information
on this.

The implementation is pretty simple. A list `Node`s `add`ed to an energy
system is stored in a special group of that energy system. The group
storing the list is stored with the energy system's `add` method being
the groups key, because it's highly unlikely that this key collides with
a key somebody else might use for a group and it avoids having to use
another "private" attribute. If this becomes a problem, its really easy
to refactor the implementation to use a dedicated attribute instead.
Whenever `groups` is accessed, the list of ungrouped `Node`s is
processed with the energy system's `groupings` and cleared afterwards,
before returning the new `groups`.
With this implementation, one can construct a bunch of `Node`s, add the
to the energy system and connect them or alter their attributes
afterwards and if any `grouping` depends on one of those modifications,
the `Node`s will be processed correctly, because the `groupings` will
only be evaluated once `groups` is accessed, at which point in time all
modification should be in place.

The only slightly complicated thing code in the implementation that
probably has to be explained is the `deque(iterable, maxlen=0)` recipe.
This just consumes the whole `iterable` without storing the result. This
is done here because we only need the side effect of modifying the
`groups`. The recipe is the fastest way of consuming an `iterable` on
CPython and it has minimal memory overhead since nothing is stored.
See the [`consume`][0] recipe in the [`itertools` documentation][1].

[0]: https://docs.python.org/3.5/library/itertools.html#itertools-recipes
[1]: https://docs.python.org/3.5/library/itertools.html