Test the #576 case that laziness can't fix
If an `Edge` which connects two `Node`s which are already grouped, i.e.
they were added to the energy system and `groups` was accessed after
their creation but before the `Edge` connecting them is created, the
connection established by the `Edge` and any attributes, properties or
other characteristics the `Edge` has, are not picked up a subsequent
evaluation of a flow grouping.
This happens because these special groupings essentially don't do
anything for `Edge`s, which is because the groupings are acting on
`inputs` and `outputs` which are intentionally empty for `Edge`s. So
when the `Node`s are grouped, the `Edge` doesn't exist yet and when the
`Edge` is grouped, nothing happens.
The code in this commit just tests that case explicitly to make sure we
have a fix in place that goes beyond just using laziness.