Start using blinker signals
This enables us to avoid the tight coupling of `Node`s and
`EnergySystem`s which we would get by wiring certain behaviour into an
undocumented special attribute like `subnodes`. It's also more flexible
because other interested objects can connect to the signal to and last
but not least we might find more interesting uses for other signals in
the future.
[Blinker][0] is actually a really nice and easy to use library.

[0]: https://pythonhosted.org/blinker/