Add a UserWarning for typical problems to facilitate debugging
A UserWarning should be raised for untypical uses even though this kind
of usage is valid if you really know what you are doing, whereas in 90%
of the cases it would be a mistake. This will help users to debug their
code but can be turned of for experienced users.