Remove tautology check in shared_limit
There was a check in shared_limit, that expressions
are not added to the model, if they are always true.
This looks like a performance optimisation, however,
I found only constructed edge cases where this was used.
Thus, it is better to potentially add useless constraints
instead of checking every time.