Workaround SolverStatus problem
There seems to be an (undocumented) change to Pyomo's API.
(They never provide a changelog, so one might argue,
every change is undocumented.) This commit allows to live
with the issue.

Testing would require using different versions on Pyomo,
and it would require testing to find out which version
introduced the change.

Fix #709