Introduce parameter "expected" for Flows
This value can be used as a hint to the optimiser to speed up
minimisation. The fature is actually a reintroduction:
Up to oemof v0.3, it was possible to have a time series "actual_value"
plus a boolean parameter "fixed". When fixed was false, it was just
taken as the initial value for optimisation of the energy system.
When compared to this previous implemenataion, having either fix=(series)
or expected=(series) might be easier to understand.