Rename Transformer to Converter
When people hear "transformer", they think of electrical devices.
However, the Transformer is neither meant to be (only) electrical
nor bidirectional (as electrical transformers typically are).
Thus, I got the suggestion to rename it to Converter.
The term is more general and does not imply that back-conversion
is possible.

Fix #728