Start moving constraint generation to factories
NonConvexFlowBlock and NonConvexInvestFlowBlock had a lot
of shared code in in their respective _create_constraints
functions. To remove duplication and to reduce complexity,
Common constraints are now generated by common factory functions.

Due to the two-step cretion/build for constraints to positive or
negative gradients (de facto steps, not gradients), these could
not be included the same way.