Simplify constraint tests
Pyomo 6.6 offers a function lp_diff(...) to unify the format
of LP files. This commit removes custom written code to do the
same. On the other hand, the LP files do not have a reproducable
order of the lines. Thus, this commit changes from line-by-line
comparission to a comparission of the sorted lines.