Fix "Flow" does not have "_in_edges"
In old versions of oemof.network, Edge was subclass to Node.
Thus, the Flow had the attribute "_in_edges" though there was
no Edge connected to an Edge. Now, the Edge (sand thus Flow)
is cleaned up in oemof.network. As attributes starting with "_"
are ignored  anyway, this commit does not even change any behaviour.