Allow sinks/sources to have multiple inputs/outputs
There was a warning and a comment that having multiple inputs/outputs
for one Sink/Source is unintended but should be possible.
I think we can remove that "strong reccomendation" as it just works.

To not let beginners fall into the trap of having a dangling Sink
or Source, like the old warning also might have prevented, inputs and
outputs are now explicitly named arguments without a default.