Fix loss calculation for non-hourly time steps
This one first slipped through due to too many irrelevant changes
in the LP files. I now added a unit test that is independent of lp files.

Note that the formula is more complicated but it just changes a value
in the lp file, so it will not significantly increase computational time.